+++ docs/source/changelog.rst
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 `Unreleased`_
 -------------
 
+`0.3.0`_ -- 2018-07-25
+----------------------
+
 Added
 ^^^^^
 
@@ -55,7 +58,8 @@ Changed
 * Initial release.
 
 
-.. _Unreleased: https://github.com/Bogdanp/molten/compare/v0.2.1...HEAD
+.. _Unreleased: https://github.com/Bogdanp/molten/compare/v0.3.0...HEAD
+.. _0.3.0: https://github.com/Bogdanp/molten/compare/v0.2.1...v0.3.0
 .. _0.2.1: https://github.com/Bogdanp/molten/compare/v0.2.0...v0.2.1
 .. _0.2.0: https://github.com/Bogdanp/molten/compare/v0.1.0...v0.2.0
 .. _0.1.0: https://github.com/Bogdanp/molten/compare/02c76ae...v0.1.0